pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Comand line utilite app for showing difference between files"
authors = ["denbon05 <denisamigoua@gmail.com>"]
license = "MIT"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
isort = "^5.8.0"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
